// src/app/enums/zone.enum.ts
export enum ZoneClassName {
    top = 'top',
    center = 'center',
    bottom = 'bottom'
}

export enum ZoneColumnType {
    topLeft = 10,
    topMiddle = 11,
    topRight = 12,
    centerLeft = 20,
    centerMiddle = 21,
    centerRight = 22,
    bottomLeft = 30,
    bottomMiddle = 31,
    bottomRight = 32
}

export enum PositionClass {
    topLeft = 'toast-top-left',
    topMiddle = 'toast-top-center',
    topRight = 'toast-top-right',
    topFull = 'toast-top-full-width',
    centerLeft = 'toast-center-left',
    centerMiddle = 'toast-center-center',
    centerRight = 'toast-center-right',
    centerFull = 'toast-center-full-width',
    bottomLeft = 'toast-bottom-left',
    bottomMiddle = 'toast-bottom-center',
    bottomRight = 'toast-bottom-right',
    bottomFull = 'toast-bottom-full-width',
}